requirements.txt
playwright>=1.49.0
apscheduler==3.10.4
python-dotenv==1.0.1
anthropic>=0.40.0